going to move forward pretending everything is an int and will then move from there once things are working

+++ src/main/java/org/launchcode/models/DateCreated.java
@@ -5,7 +5,7 @@ package org.launchcode.models;
  */
 public class DateCreated extends BookField {
 
-    public DateCreated(Float value) {
+    public DateCreated(int value) {
         super(value);
     }
 }
